backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
firebase-admin==6.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1

sentence-transformers==2.2.2
transformers==4.30.2
huggingface_hub==0.16.4

numpy==1.26.4
werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
faiss-cpu==1.8.0
python-dateutil==2.8.2
